RAD-747 Support for conditional files
