+++ src/test/resources/sample/src/main/resources/templates/__projectName__-model/src/main/java/__packagePath__/model/ExampleOptional.java__IF_isSet__
@@ -1,0 +1,29 @@
+package @packageName@.model;
+
+public class ExampleOptional {
+
+  String name;
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  @postmapping
+  public ResponseEntity createProduct(@requestbody @Valid ProductDetail product) {
+    log.info("REST[createProduct] Rx[{}]", toStringHelper.asString(product));
+
+    ProductDetail productDetail = productDetailService.create(product);
+
+    log.info("REST[createProduct] Tx[{}]", toStringHelper.asString(productDetail));
+    return new ResponseEntity<>(productDetail, HttpStatus.OK);
+  }
+
+  @Override
+  public String toString() {
+    return name;
+  }
+}
